[nginx] re-organize Chopped::Nginx into nice little files
