+++ site-cookbooks/chopped_nginx/libraries/chopped.rb
@@ -1,0 +1,6 @@
+module Chopped
+end
+
+require_relative './chopped/block_property_support'
+require_relative './chopped/combined_scope'
+require_relative './chopped/nginx'
