+++ site-cookbooks/chopped_nginx/libraries/chopped/nginx/directives_list.txt
